fix(core): :bug: border documentation ui related changes
border documentation ui related changes in guide/styles

ref: #115
